Combine previous packages into one package
