--- projects/core/README.md
@@ -1,36 +1,0 @@
-<p align="center">
-  <img height="200px" width="200px" style="text-align: center;" src="https://cdn.rawgit.com/MurhafSousli/ngx-progressbar/master/src/assets/logo.svg">
-  <h1 align="center">Angular Progressbar</h1>
-</p>
-
-A nanoscopic progress bar. Featuring realistic trickle animations to convince your users that something is happening!
-
-___
-[![npm](https://img.shields.io/badge/demo-online-ed1c46.svg)](https://murhafsousli.github.io/ngx-progressbar/)
-[![npm](https://img.shields.io/badge/stackblitz-online-orange.svg)](https://stackblitz.com/edit/ngx-progressbar)
-[![npm](https://img.shields.io/npm/v/@ngx-progressbar/core.svg?maxAge=2592000?style=plastic)](https://www.npmjs.com/package/@ngx-progressbar/core) 
-[![Build Status](https://travis-ci.org/MurhafSousli/ngx-progressbar.svg?branch=master)](https://travis-ci.org/MurhafSousli/ngx-progressbar)
-[![npm](https://img.shields.io/npm/dt/@ngx-progressbar/core.svg?maxAge=2592000?style=plastic)](https://www.npmjs.com/package/@ngx-progressbar/core)
-[![npm](https://img.shields.io/npm/dm/@ngx-progressbar/core.svg)](https://www.npmjs.com/package/@ngx-progressbar/core)
-[![npm bundle size (minified + gzip)](https://img.shields.io/bundlephobia/minzip/@ngx-progressbar/core.svg)](https://bundlephobia.com/result?p=@ngx-progressbar/core)
-[![npm](https://img.shields.io/npm/l/express.svg?maxAge=2592000)](https://github.com/MurhafSousli/ngx-progressbar/blob/master/LICENSE)
-
-
-### The documentation is available at the [wiki page](https://github.com/MurhafSousli/ngx-progressbar/wiki)
-
-## Support
-
-Please give NgProgress a :star:
-
-[![npm](https://c5.patreon.com/external/logo/become_a_patron_button.png)](https://www.patreon.com/bePatron?u=5594898)
-
-## Issues
-
-If you identify any errors in the library, or have an idea for an improvement, please open an [issue](https://github.com/MurhafSousli/ngx-progressbar/issues).
-
-## Author
-
- **[Murhaf Sousli](http://murhafsousli.com)**
-
-- [github/murhafsousli](https://github.com/MurhafSousli)
-- [twitter/murhafsousli](https://twitter.com/MurhafSousli)
