--- projects/core/src/public_api.ts
@@ -1,5 +1,0 @@
-export * from './lib/ng-progress.module';
-export * from './lib/ng-progress.component';
-export * from './lib/ng-progress-ref';
-export * from './lib/ng-progress.service';
-export * from './lib/ng-progress.interface';
