--- projects/http/ng-package.json
@@ -1,11 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/http",
-  "deleteDestPath": false,
-  "lib": {
-    "entryFile": "src/public_api.ts",
-    "externals": {
-      "@ngx-progressbar/core": "@ngx-progressbar/core"
-    }
-  }
-}
