--- projects/http/ng-package.prod.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/http",
-  "lib": {
-    "entryFile": "src/public_api.ts",
-    "externals": {
-      "@ngx-progressbar/core": "@ngx-progressbar/core"
-    }
-  }
-}
