--- projects/http/package.json
@@ -1,39 +1,0 @@
-{
-  "name": "@ngx-progressbar/http",
-  "version": "VERSION",
-  "description": "A nanoscopic progress bar. Featuring realistic trickle animations to convince your users that something is happening!",
-  "private": false,
-  "homepage": "http://github.com/murhafsousli/ngx-progressbar",
-  "author": {
-    "name": "Murhaf Sousli",
-    "url": "http://github.com/murhafsousli"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/murhafsousli/ngx-progressbar.git"
-  },
-  "keywords": [
-    "angular",
-    "progress",
-    "progressbar",
-    "loading",
-    "http",
-    "xhr",
-    "interceptors",
-    "loader",
-    "slim",
-    "youtube",
-    "medium",
-    "github"
-  ],
-  "bugs": {
-    "url": "http://github.com/murhafsousli/ngx-progressbar/issues"
-  },
-  "license": "MIT",
-  "peerDependencies": {
-    "@angular/common": ">=6.0.0",
-    "@angular/core": ">=6.0.0",
-    "rxjs": ">=6.0.0",
-    "@ngx-progressbar/core": "VERSION"
-  }
-}
