+++ projects/ngx-progressbar/http/package.json
@@ -1,0 +1,3 @@
+{
+  "ngPackage": {}
+}
