+++ projects/ngx-progressbar/http/src/ng-progress-http.interface.ts
