+++ projects/ngx-progressbar/http/src/ng-progress-http.module.ts
