+++ projects/ngx-progressbar/http/src/public_api.ts
@@ -1,0 +1,2 @@
+export * from './ng-progress-http.module';
+export * from './ng-progress-http.interface';
