+++ projects/ngx-progressbar/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "ngx-progressbar",
+  "version": "6.0.0-beta.0",
+  "peerDependencies": {
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@angular/router": "^8.0.0"
+  }
+}
