+++ projects/ngx-progressbar/router/package.json
@@ -1,0 +1,3 @@
+{
+  "ngPackage": {}
+}
