+++ projects/ngx-progressbar/router/src/ng-progress-router.interface.ts
