+++ projects/ngx-progressbar/router/src/ng-progress-router.module.ts
