+++ projects/ngx-progressbar/router/src/public_api.ts
@@ -1,0 +1,2 @@
+export * from './ng-progress-router.module';
+export * from './ng-progress-router.interface';
