+++ projects/ngx-progressbar/src/lib/ng-progress.component.scss
@@ -1,0 +1,151 @@
+ng-progress {
+  z-index: 999999;
+  pointer-events: none;
+
+  &[fixed='true'] {
+    .ng-progress-bar, .ng-spinner {
+      position: fixed;
+    }
+    .ng-spinner {
+      top: 15px;
+    }
+    &[spinnerPosition='left'] .ng-spinner {
+      left: 15px;
+    }
+    &[spinnerPosition='right'] .ng-spinner {
+      right: 15px;
+    }
+  }
+
+  &[thick='true'] {
+    .ng-spinner-icon {
+      width: 24px;
+      height: 24px;
+      border-width: 3px;
+    }
+    .ng-bar-placeholder {
+      height: 3px !important;
+    }
+  }
+
+  &[dir='ltr+'], &[dir='ltr-'] {
+    .ng-meteor {
+      transform: rotate(3deg);
+    }
+    &[thick='true'] .ng-meteor {
+      transform: rotate(4deg);
+    }
+  }
+
+  &[dir='ltr+'], &[dir='rtl+'] {
+    .ng-bar {
+      margin-left: -100%;
+    }
+    .ng-meteor {
+      right: 0;
+    }
+  }
+
+  &[dir='ltr+'], &[dir='rtl-'] {
+    .ng-meteor {
+      top: -3px;
+    }
+    &[thick='true'] .ng-meteor {
+      top: -4px;
+    }
+  }
+
+  &[dir='ltr-'], &[dir='rtl+'] {
+    .ng-meteor {
+      bottom: -3px;
+    }
+    &[thick='true'] .ng-meteor {
+      bottom: -4px;
+    }
+    // reverse dir
+    .ng-bar-placeholder {
+      transform: rotate(180deg);
+    }
+    .ng-spinner-icon {
+      animation-direction: reverse;
+    }
+  }
+
+  &[dir='rtl+'], &[dir='rtl-'] {
+    .ng-meteor {
+      transform: rotate(-3deg);
+    }
+    &[thick='true'] .ng-meteor {
+      transform: rotate(-4deg);
+    }
+  }
+
+  &[spinnerPosition='left'] .ng-spinner {
+    left: 10px;
+  }
+
+  &[spinnerPosition='right'] .ng-spinner {
+    right: 10px;
+  }
+}
+
+.ng-progress-bar {
+  position: relative;
+  z-index: 999999;
+  top: 0;
+  left: 0;
+  width: 100%;
+  zoom: 1;
+  filter: alpha(opacity=0);
+  opacity: 0;
+  &[active="true"] {
+    filter: alpha(opacity=100);
+    opacity: 1;
+    transition: none;
+  }
+}
+
+.ng-bar-placeholder {
+  position: absolute;
+  height: 2px;
+  width: 100%;
+}
+
+.ng-bar {
+  width: 100%;
+  height: 100%;
+  transform: translate(-100%, 0, 0);
+}
+
+.ng-meteor {
+  display: block;
+  position: absolute;
+  width: 100px;
+  height: 100%;
+  opacity: 1;
+}
+
+.ng-spinner {
+  position: absolute;
+  display: block;
+  z-index: 1031;
+  top: 10px;
+}
+
+.ng-spinner-icon {
+  width: 18px;
+  height: 18px;
+  box-sizing: border-box;
+  animation: spinner-animation 250ms linear infinite;
+  border: 2px solid transparent;
+  border-radius: 50%;
+}
+
+@keyframes spinner-animation {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
