+++ projects/ngx-progressbar/src/lib/ng-progress.interface.ts
