+++ projects/ngx-progressbar/src/lib/ng-progress.module.ts
