+++ projects/ngx-progressbar/src/lib/ng-progress.service.ts
