+++ projects/ngx-progressbar/src/public-api.ts
@@ -1,0 +1,9 @@
+/*
+ * Public API Surface of ngx-progressbar
+ */
+
+export * from './lib/ng-progress-ref';
+export * from './lib/ng-progress.component';
+export * from './lib/ng-progress.module';
+export * from './lib/ng-progress.service';
+export * from './lib/ng-progress.interface';
