+++ projects/ngx-progressbar/tsconfig.spec.json
