--- projects/router/src/public_api.ts
@@ -1,1 +1,0 @@
-export * from './lib/ng-progress-router.module';
