--- projects/router/tslint.json
@@ -1,17 +1,0 @@
-{
-    "extends": "../../src/tslint.json",
-    "rules": {
-        "directive-selector": [
-            true,
-            "attribute",
-            "lib",
-            "camelCase"
-        ],
-        "component-selector": [
-            true,
-            "element",
-            "lib",
-            "kebab-case"
-        ]
-    }
-}
